Entregable/requirements.txt
numpy==1.26.4
packaging==24.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
urllib3==2.2.1
